Complete user routes with middlewares (tasks OT290-34,36,37)

+++ controllers/users.js
@@ -44,24 +44,26 @@ const createUser = async (req, res) => {
 
 const updateUser = async (req, res) => {
   try {
-    const { id } = req.params;
+    const { id, userId } = req.params;
     const { firstName, lastName, email, password, photo, roleId } = req.body;
-
     const user = await User.findByPk(id);
     if (!user) {
       return res.sendStatus(404);
     }
-
-    const update = await User.update(
-      { firstName, lastName, email, password, photo, roleId },
-      {
-        where: {
-          id,
-        },
-      }
-    );
-    const userUpdated = await User.findByPk(id);
-    return res.status(200).json(userUpdated);
+    if (id == userId) {
+      const update = await User.update(
+        { firstName, lastName, email, password, photo, roleId },
+        {
+          where: {
+            id,
+          },
+        }
+      );
+      const userUpdated = await User.findByPk(id);
+      return res.status(200).json(userUpdated);
+    } else {
+      return res.status(401).json(error);
+    }
   } catch (error) {
     return res.status(500).json(error);
   }
@@ -69,23 +71,27 @@ const updateUser = async (req, res) => {
 
 const deleteUser = async (req, res) => {
   try {
-    const { id } = req.params;
+    const { id, userId } = req.params;
     const user = await User.findByPk(id);
     if (!user) {
       return res.sendStatus(404);
     }
-    await User.destroy({
-      where: {
-        id,
-      },
-    });
-    const userDeleted = await User.findByPk(id, {
-      paranoid: false,
-    });
-    return res.status(200).json(userDeleted);
+    if (id == userId) {
+      await User.destroy({
+        where: {
+          id,
+        },
+      });
+      const userDeleted = await User.findByPk(id, {
+        paranoid: false,
+      });
+      return res.status(200).json(userDeleted);
+    } else {
+      return res.status(401).json(error);
+    }
   } catch (error) {
     return res.status(500).json(error);
-  }
+  }    
 };
 
 module.exports = {
